project: Apply more C++ paradigms to the code
- Use auto in places where code clarity is improved or identical.
- Replace trivial constructors and destructors with default.
- Use true random for random generation.
- Use std::string_view where it is valid to do so.
- Apply const where it is valid to do so.
- Use references where it is valid to do so.
- Manually optimize memory usage with std::move and std::copy.
- Opt for memory efficient containers where the size is known ahead of time.

+++ source/obs/gs/gs-mipmapper.cpp
@@ -327,11 +327,11 @@ void streamfx::obs::gs::mipmapper::rebuild(std::shared_ptr<streamfx::obs::gs::te
 			// Copy from the render target to the target mip level.
 #ifdef _WIN32
 			if (gs_get_device_type() == GS_DEVICE_DIRECT3D_11) {
-				d3d_copy_subregion(d3dinfo, _rt->get_texture(), mip, cwidth, cheight);
+				d3d_copy_subregion(d3dinfo, _rt->get_texture(), static_cast<uint32_t>(mip), cwidth, cheight);
 			}
 #endif
 			if (gs_get_device_type() == GS_DEVICE_OPENGL) {
-				opengl_copy_subregion(oglinfo, _rt->get_texture(), mip, cwidth, cheight);
+				opengl_copy_subregion(oglinfo, _rt->get_texture(), static_cast<uint32_t>(mip), cwidth, cheight);
 			}
 		}
